+++ src/Utilities/Communication/comm_exchange_sizes.c
@@ -58,14 +58,14 @@ extern "C" {
 /*****************************************************************************/
 
 int       Zoltan_Comm_Exchange_Sizes(
-int      *sizes_to,		/* value I need to exchange (size of true msg) */
+size_t      *sizes_to,		/* value I need to exchange (size of true msg) */
 int      *procs_to,		/* procs I send to */
 int       nsends,		/* number of messages I'll send */
 int       self_msg,		/* do I copy data to myself? */
-int      *sizes_from,		/* (returned) size of all my receives */
+size_t      *sizes_from,		/* (returned) size of all my receives */
 int      *procs_from,		/* procs I recv from */
 int       nrecvs,		/* number of messages I receive */
-int      *total_recv_size,	/* (returned) sum of all incoming sizes */
+size_t      *total_recv_size,	/* (returned) sum of all incoming sizes */
 int       my_proc,		/* my processor number */
 int       tag,			/* message tag I can use */
 MPI_Comm  comm) {		/* communicator */
@@ -80,7 +80,7 @@ MPI_Comm  comm) {		/* communicator */
     self_index_to = -1;
     for (i = 0; i < nsends + self_msg; i++) {
 	if (procs_to[i] != my_proc) 
-	    MPI_Send((void *) &sizes_to[i], 1, MPI_INT, procs_to[i], tag, comm);
+	    MPI_Send((void *) &sizes_to[i], 1, MPI_UNSIGNED_LONG, procs_to[i], tag, comm);
 	else
 	    self_index_to = i;
     }
@@ -88,7 +88,7 @@ MPI_Comm  comm) {		/* communicator */
     *total_recv_size = 0;
     for (i = 0; i < nrecvs + self_msg; i++) {
 	if (procs_from[i] != my_proc) 
-	    MPI_Recv((void *) &sizes_from[i], 1, MPI_INT, procs_from[i],
+	    MPI_Recv((void *) &sizes_from[i], 1, MPI_UNSIGNED_LONG, procs_from[i],
 		     tag, comm, &status);
 	else 
 	    sizes_from[i] = sizes_to[self_index_to];
